backend/requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.0
backcall==0.2.0
click==8.1.3
decorator==5.1.1
Django==4.1.6
django-cors-headers==3.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
Faker==15.3.4
Flask==2.2.2
Flask-Cors==3.0.10
gunicorn==20.1.0
importlib-metadata==5.1.0
install==1.3.5
ipdb==0.13.9
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
numpy==1.22.3
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.33
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
toml==0.10.2
traitlets==5.6.0
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.11.0
